+++ Game/Game/BlockAir.hpp
@@ -8,7 +8,7 @@
  */
 class BlockAir : public Block {
 public:
-	BlockAir(int width, int height);
+	BlockAir(double width, double height);
 	/**
 	 * 描画時に呼ばれる．
 	 */
